+++ src/adapters/openweathermap.adapter.js
@@ -1,7 +1,7 @@
 import axios from 'axios';
 import openweathermapConfig from '../config/openweathermap.json';
 
-const getWeather = async () => {
+export const getWeather = async () => {
     try {
         const resp = await axios.get(`${openweathermapConfig.host}/weather?q=${openweathermapConfig.city}&appid=${openweathermapConfig.key}&units=metric`);
         return (({name, weather, main, wind}) => ({name, weather, main, wind}))(resp.data);
@@ -10,7 +10,7 @@ const getWeather = async () => {
     }
 }
 
-const getForecast = async () => {
+export const getForecast = async () => {
     try {
         const resp = await axios.get(`${openweathermapConfig.host}/forecast?q=${openweathermapConfig.city}&appid=${openweathermapConfig.key}&units=metric`);
         return resp.data.list.reduce((acc, curr, i) => (i + 1) % 8 === 0 ? [...acc, (({dt, weather, main, wind}) => ({dt, weather, main, wind}))(curr)] : acc, []);
@@ -18,5 +18,3 @@ const getForecast = async () => {
         throw err;
     }
 }
-
-export {getWeather, getForecast};
