+++ src/components/City/city.scss
@@ -88,6 +88,7 @@ $scaling_distance: calc(-25px + 4.5vw);
     padding: 0;
     line-height: 0;
     border-radius: 8px;
+    border: none;
     height: 1.2em;
     width: 1.2em;
     margin: 0.125em 0;
